+++ gobin/jwt.go
@@ -2,24 +2,10 @@ package gobin
 
 import (
 	"context"
-	"errors"
-	"fmt"
 	"net/http"
-	"strings"
 	"time"
 
-	"github.com/go-chi/chi/v5"
 	"github.com/go-jose/go-jose/v3/jwt"
-)
-
-var (
-	ErrNoPermissions     = errors.New("no permissions provided")
-	ErrUnknownPermission = func(p Permission) error {
-		return fmt.Errorf("unknown permission: %s", p)
-	}
-	ErrPermissionDenied = func(p Permission) error {
-		return fmt.Errorf("permission denied: %s", p)
-	}
 )
 
 type Permission string
@@ -51,37 +37,12 @@ type claimsKey struct{}
 
 var claimsContextKey = claimsKey{}
 
-func (s *Server) JWTMiddleware(next http.Handler) http.Handler {
-	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		tokenString := r.Header.Get("Authorization")
-		if len(tokenString) > 7 && strings.ToUpper(tokenString[0:6]) == "BEARER" {
-			tokenString = tokenString[7:]
-		}
-
-		var claims Claims
-		if tokenString == "" {
-			documentID := chi.URLParam(r, "documentID")
-			claims = newClaims(documentID, nil)
-		} else {
-			token, err := jwt.ParseSigned(tokenString)
-			if err != nil {
-				s.error(w, r, err, http.StatusUnauthorized)
-				return
-			}
-
-			if err = token.Claims([]byte(s.cfg.JWTSecret), &claims); err != nil {
-				s.error(w, r, err, http.StatusUnauthorized)
-				return
-			}
-		}
-
-		ctx := context.WithValue(r.Context(), claimsContextKey, claims)
-		next.ServeHTTP(w, r.WithContext(ctx))
-	})
-}
-
 func GetClaims(r *http.Request) Claims {
 	return r.Context().Value(claimsContextKey).(Claims)
+}
+
+func SetClaims(r *http.Request, claims Claims) *http.Request {
+	return r.WithContext(context.WithValue(r.Context(), claimsContextKey, claims))
 }
 
 func (s *Server) NewToken(documentID string, permissions []Permission) (string, error) {
@@ -99,10 +60,6 @@ func newClaims(documentID string, permissions []Permission) Claims {
 	}
 }
 
-func GetWebhookSecret(r *http.Request) string {
-	secretStr := r.Header.Get("Authorization")
-	if len(secretStr) > 7 && strings.ToUpper(secretStr[0:6]) == "SECRET" {
-		return secretStr[7:]
-	}
-	return ""
+func EmptyClaims(documentID string) Claims {
+	return newClaims(documentID, nil)
 }
